chore: (litmus-portal) Refactoring and bug fixes (#2027)
This commit has the following changes:
- folder structure change for models and useEffect fixes
- user redux fixed
- graphql documents re-organised
