+++ litmus-portal/frontend/src/components/Sections/ChaosWorkflows/BrowseWorkflow/index.tsx
@@ -22,7 +22,7 @@ import {
   WorkflowDataVars,
   WorkflowRun,
   WorkflowSubscription,
-} from '../../../../models/workflowData';
+} from '../../../../models/graphql/workflowData';
 import { RootState } from '../../../../redux/reducers';
 import {
   sortAlphaAsc,
