+++ litmus-portal/frontend/src/containers/app/App.tsx
@@ -2,7 +2,7 @@ import React, { lazy, Suspense, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { Redirect, Route, Router, Switch } from 'react-router-dom';
 import Loader from '../../components/Loader';
-import { UserData } from '../../models/user';
+import { UserData } from '../../models/redux/user';
 import useActions from '../../redux/actions';
 import * as AnalyticsActions from '../../redux/actions/analytics';
 import { history } from '../../redux/configureStore';
