+++ litmus-portal/frontend/src/models/redux/user.ts
@@ -1,0 +1,26 @@
+export interface UpdateUser {
+  selectedProjectID: string;
+  selectedProjectName: string;
+  userRole: string;
+}
+
+export interface UserData extends UpdateUser {
+  username: string;
+  token: string;
+  exp: number;
+}
+export enum UserActions {
+  LOAD_USER_DETAILS = 'LOAD_USER_DETAILS',
+  UPDATE_USER_DETAILS = 'UPDATE_USER_DETAILS',
+  LOGOUT_USER = 'LOGOUT_USER',
+}
+
+interface UserActionType<T, P> {
+  type: T;
+  payload: P;
+}
+
+export type UserAction =
+  | UserActionType<typeof UserActions.LOAD_USER_DETAILS, string>
+  | UserActionType<typeof UserActions.LOGOUT_USER, string>
+  | UserActionType<typeof UserActions.UPDATE_USER_DETAILS, UpdateUser>;
