+++ litmus-portal/frontend/src/redux/actions/nodeSelection.ts
@@ -1,9 +1,9 @@
 /* eslint-disable import/prefer-default-export */
+import { Node } from '../../models/graphql/workflowData';
 import {
   NodeSelectionAction,
   NodeSelectionActions,
-} from '../../models/nodeSelection';
-import { Node } from '../../models/workflowData';
+} from '../../models/redux/nodeSelection';
 
 export function selectNode(node: Node): NodeSelectionAction {
   return {
